Improved logging and optimised functions

+++ disabling_savedsearches.py
@@ -1,17 +1,20 @@
 ### DISCLAIMER
 #
-### THE SCRIPT IS NOT PERFECT
-#
-### USE SCRIPT AT YOUR OWN RISK
-#
-### ALWAYS VERIFY RESULTS
+# THE SCRIPT IS NOT PERFECT
+# USE SCRIPT AT YOUR OWN RISK
+# ALWAYS VERIFY RESULTS
 
 import os
 import shutil
 import sys
+import logging
 
 __name__ = "disabling_savedsearches.py"
 __author__ = "Michel de Jong"
+
+# Create a global logger
+logger = logging.getLogger("script_logger")
+logger.setLevel(logging.DEBUG)
 
 # Function to process the savedsearches.conf file
 def process_file(file_path, debug=False):
@@ -19,43 +22,51 @@ def process_file(file_path, debug=False):
         # Read the savedsearches.conf file
         with open(file_path, "r") as f:
             lines = f.readlines()
+            
+        modified = False
 
         # Loop over the lines in the file
         for i in range(len(lines)):
             line = lines[i].strip()
+            
             # Check if the line starts a new stanza
             if line.startswith("[") and not line.startswith("[default]") and not line.endswith("\\"):
                 stanza_name = line[1:-1]  # Extract the stanza name without brackets
+                
                 # Check if the stanza has a disabled statement
                 has_disabled = False
                 for j in range(i+1, len(lines)):
                     next_line = lines[j].strip()
                     if next_line.startswith("["):
                         break
-                    if next_line.startswith("disabled"):
+                    if next_line.startswith("disabled ="):
                         has_disabled = True
                         # Check if the disabled statement is already false or 0
                         if "false" in next_line or "0" in next_line:
                             # Disable the stanza by changing the statement to "disabled = 1"
                             lines[j] = "disabled = 1\n"
+                            modified = True
                             if debug:
-                                print(f"DEBUG: Disabled Stanza '{stanza_name}' in {file_path}")
+                                logger.debug(f"DEBUG: Disabled Stanza '{stanza_name}' in {file_path}")
                         break
                     if next_line.startswith("search") or next_line.startswith("|"):
-                        break
+                        continue
+                    
                 # If the stanza doesn't have a disabled statement, add one
                 if not has_disabled:
                     lines.insert(i+1, "disabled = 1\n")
+                    modified = True
                     if debug:
-                        print(f"DEBUG: Inserted 'disabled = 1' for Stanza '{stanza_name}' in {file_path}")
+                        logger.debug(f"DEBUG: Inserted 'disabled = 1' for Stanza '{stanza_name}' in {file_path}")
 
-        # Write the modified savedsearches.conf file
-        with open(file_path, "w") as f:
-            f.writelines(lines)
-            if debug:
-                print(f"DEBUG: Modified {file_path}")
-            else:
-                print(f"INFO: Disabled saved searches in {file_path}")
+        # If changes were made, write the modified savedsearches.conf file
+        if modified:
+            with open(file_path, "w") as f:
+                f.writelines(lines)
+                if debug:
+                    logger.debug(f"DEBUG: Modified {file_path}")
+                else:
+                    print(f"INFO: Disabled saved searches in {file_path}")
 
     except PermissionError:
         print(f"ERROR: Permission denied: {file_path}")
@@ -72,16 +83,25 @@ if __name__ == "disabling_savedsearches.py":
     disabled_ss_dir = os.path.join(os.path.dirname(apps_dir), "apps_ss_disabled")
 
     # Copy the apps to apps_ss_disabled directory
-    while os.path.exists(disabled_ss_dir):
+    if os.path.exists(disabled_ss_dir):
         decision = input(f"Destination directory {disabled_ss_dir} already exists. Proceed and overwrite? (y/n): \n")
         if decision.lower() == "y":
             shutil.rmtree(disabled_ss_dir)
-            break
-        elif decision.lower() == "n":
+        else:
             print("Exiting the script")
             exit(0)
-    print("Copying the apps directory to apps_ss_disabled")
+            
+    print(f"Copying the apps directory to {disabled_ss_dir}")
     shutil.copytree(apps_dir, disabled_ss_dir)
+    
+    if debug_mode:
+        # Create the "_DEBUG" directory inside "apps_ss_disabled"
+        logging_dir = os.path.join(disabled_ss_dir, "_DEBUG")
+        os.makedirs(logging_dir, exist_ok=True)
+        
+        # Configure logfile for DEBUG logs
+        log_filename = os.path.join(logging_dir, "debug.log")
+        logging.basicConfig(filename=log_filename, level=logging.DEBUG, format="%(asctime)s - %(levelname)s: %(message)s")
 
     # Loop over all directories in the apps_ss_disabled directory
     for app_name in os.listdir(disabled_ss_dir):
